+++ src/main/java/ma/ens/AviCultureBackend/product/modal/dto/ProductFoodDto.java
@@ -12,7 +12,7 @@ public record ProductFoodDto (
         @NotNull(message = "storage Building can't be null")
         BuildingDto storageBuilding,
         Double unitaryPrice,
-        Integer quantity,
+        Long quantity,
         Long foodCategoryId,
 
         String remarks
